+++ biobox/container.py
@@ -19,18 +19,3 @@ def is_running(id_):
     Checks if the container is running
     """
     return metadata_lookup(['State', 'Running'], id_)
-
-def collect_runtime_metrics(container_id, interval = 15, warmup = 1):
-    """
-    Collects runtime metrics about the container at the specified intervals
-    """
-    import time
-    time.sleep(warmup)
-
-    stream = util.client().stats(container_id, decode = True, stream = True)
-
-    stats = []
-    while is_running(container_id):
-        stats.append(next(stream))
-        time.sleep(interval)
-    return stats